legacy/requirements.txt
setuptools==74.1.1
Werkzeug==3.0.4
Jinja2==3.1.4
Flask==3.0.3
bcrypt==4.2.0
python-dotenv==1.0.1